backend/requirements.txt
pydantic
fastapi[standard]
python-jose
uvicorn[standard]
python-dotenv
argon2-cffi
pynput
websockets